@@ -7,6 +7,7 @@
     "build": "react-app-rewired build",
     "test": "react-app-rewired test",
     "eject": "react-scripts eject"
+    
   },
   "dependencies": {
     "@cometchat-pro/chat": "^3.0.10",
@@ -14,11 +15,11 @@
     "@nomiclabs/hardhat-waffle": "^2.0.3",
     "ethereum-waffle": "^3.4.4",
     "ethers": "^5.6.9",
-    "hardhat": "^2.10.1",
+    "hardhat": "^2.22.3",
     "ipfs-http-client": "^57.0.3",
     "moment": "^2.29.4",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.0.2",
+    "react-dom": "^18.0.2",
     "react-hooks-global-state": "^1.0.2",
     "react-icons": "^4.3.1",
     "react-identicons": "^1.2.5",
