@@ -9,9 +9,14 @@ import { useGlobalState } from '../store'
 const Home = () => {
   const [projects] = useGlobalState('projects')
 
-  useEffect(async () => {
-    await loadProjects()
+  useEffect(() => {
+    const fetchData = async () => {
+      await loadProjects()
+    }
+
+    fetchData();
   }, [])
+  
   return (
     <>
       <Hero />
